[FLINK-13972] Move PreviewPlanEnv to flink-tests
